+++ src/main/kotlin/lotto/domain/LottoAgent.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class LottoAgent(val exchange: Exchange<LottoNumbers>) {
     constructor(lottoDrawMachine: LottoDrawMachine) : this(Exchange.Lotto(lottoDrawMachine))
