+++ src/main/kotlin/lotto/domain/LottoDrawMachine.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class LottoDrawMachine(private val pool: Set<Int>) {
     val size: Int = pool.size
