+++ src/main/kotlin/lotto/domain/LottoTicket.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 data class LottoTicket(val count: Int) {
     constructor(money: Money) : this(money / TicketAmount().money)
