+++ src/main/kotlin/lotto/domain/Match.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 data class Match(val count: Int = 0) {
     constructor(left: List<Int>, right: List<Int>) : this(left.intersect(right).size)
