+++ src/main/kotlin/lotto/domain/WinningNumbers.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class WinningNumbers(private val winningNumbers: LottoNumbers, val bonusNumber: Int) {
     fun match(lottoNumbers: LottoNumbers): Match = Match(winningNumbers, lottoNumbers)
