+++ src/test/kotlin/lotto/domain/LottoAgentTest.kt
@@ -1,5 +1,10 @@
-package lotto
+package lotto.domain
 
+import lotto.domain.Exchange
+import lotto.domain.LottoAgent
+import lotto.domain.LottoDrawMachine
+import lotto.domain.LottoNumbers
+import lotto.domain.Money
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
