+++ src/test/kotlin/lotto/domain/LottoGameTest.kt
@@ -1,7 +1,7 @@
-package lotto
+package lotto.domain
 
-import lotto.Ranking.Rank.FIRST
-import lotto.Ranking.Rank.MISS
+import lotto.domain.Ranking.Rank.FIRST
+import lotto.domain.Ranking.Rank.MISS
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
