+++ src/test/kotlin/lotto/domain/LottoTicketTest.kt
@@ -1,5 +1,7 @@
-package lotto
+package lotto.domain
 
+import lotto.domain.LottoTicket
+import lotto.domain.Money
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
