+++ src/test/kotlin/lotto/domain/WinningNumbersTest.kt
@@ -1,5 +1,7 @@
-package lotto
+package lotto.domain
 
+import lotto.domain.LottoNumbers
+import lotto.domain.WinningNumbers
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
